# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: orDough Bakery API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, ready, delivered, cancelled]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # --- INGREDIENTS ---
...
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List all recipes
      responses:
        "200":
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeBody"
      responses:
        "201":
          description: Created recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
  /recipes/{id}:
    patch:
      operationId: updateRecipe
      tags: [recipes]
      summary: Update a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecipeBody"
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Not found
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Delete a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # --- REVIEWS ---
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List all reviews
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReviewBody"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          description: Not found
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # --- SETTINGS ---
...
components:
  schemas:
...
    OrderItem:
      type: object
      properties:
        menuItemId:
          type: ["string", "null"]
        name:
          type: string
        emoji:
          type: string
        qty:
          type: integer
        price:
          type: number
      required:
        - name
        - emoji
        - qty
        - price
    Order:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: ["string", "null"]
        customerName:
          type: string
        date:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        discount:
          type: number
        deliveryCharge:
          type: number
        status:
          type: string
          enum: [pending, ready, delivered, cancelled]
        notes:
          type: ["string", "null"]
        invoiceNumber:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - date
        - items
        - discount
        - deliveryCharge
        - status
        - invoiceNumber
        - createdAt
    CreateOrderBody:
      type: object
      properties:
        customerId:
          type: ["string", "null"]
        customerName:
          type: string
        date:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        discount:
          type: number
        deliveryCharge:
          type: number
        status:
          type: string
          enum: [pending, ready, delivered, cancelled]
        notes:
          type: ["string", "null"]
      required:
        - customerName
        - items
    UpdateOrderBody:
      type: object
      properties:
        customerId:
          type: ["string", "null"]
        customerName:
          type: string
        date:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        discount:
          type: number
        deliveryCharge:
          type: number
        status:
          type: string
          enum: [pending, ready, delivered, cancelled]
        notes:
          type: ["string", "null"]
...
    RecipeIngredient:
      type: object
      properties:
        ingredientId:
          type: string
        ingredientName:
          type: string
        quantity:
          type: number
        unit:
          type: string
      required:
        - ingredientId
        - ingredientName
        - quantity
        - unit
    Recipe:
      type: object
      properties:
        id:
          type: string
        menuItemId:
          type: ["string", "null"]
        menuItemName:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        instructions:
          type: ["string", "null"]
        servings:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - menuItemName
        - ingredients
        - servings
        - createdAt
    CreateRecipeBody:
      type: object
      properties:
        menuItemId:
          type: ["string", "null"]
        menuItemName:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        instructions:
          type: ["string", "null"]
        servings:
          type: integer
      required:
        - menuItemName
        - ingredients
        - servings
    UpdateRecipeBody:
      type: object
      properties:
        menuItemId:
          type: ["string", "null"]
        menuItemName:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        instructions:
          type: ["string", "null"]
        servings:
          type: integer
    Review:
      type: object
      properties:
        id:
          type: string
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        product:
          type: ["string", "null"]
        approved:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - rating
        - comment
        - approved
        - createdAt
    CreateReviewBody:
      type: object
      properties:
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        product:
          type: ["string", "null"]
        approved:
          type: boolean
      required:
        - customerName
        - rating
        - comment
    UpdateReviewBody:
      type: object
      properties:
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        product:
          type: ["string", "null"]
        approved:
          type: boolean
...