# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * orDough Bakery API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyStatus } from './createOrderBodyStatus';
import type { OrderItem } from './orderItem';

export interface CreateOrderBody {
  /** @nullable */
  customerId?: string | null;
  customerName: string;
  date?: Date;
  items: OrderItem[];
  discount?: number;
  deliveryCharge?: number;
  status?: CreateOrderBodyStatus;
  /** @nullable */
  notes?: string | null;
}